rcs2git.c:latin1_alpha(): make the argument int
latin1_alpha() is declared as taking unsigned char; however, it is actually
given an int type argument which can, in particular, be EOF == (int) -1.  EOF
is then converted to (unsigned char) 255 which latin1_alpha() happily accepts
as valid.

As a result, the inner loop in expandline() for KDELIM case, where keyword is
consumed from the input, doesn't stop on EOF.  Therefore in the (corner) case
where the CVS file ends in $smth with no newline, on output we get
$smth\xff\xff\xff\xff (i.e. filled with \xff up to KEYLENGTH == 8 characters
after the $).

Declaring latin1_alpha() as taking int fixes the problem.

+++ rcs2git.c
@@ -270,7 +270,7 @@ static void out_awrite(char const *s, size_t len)
 		out_putc(*s++);
 }
 
-static int latin1_alpha(uchar c)
+static int latin1_alpha(int c)
 {
 	if (c >= 192 && c != 215 && c != 247 ) return 1;
 	if ((c >= 97 && c < 123) || (c >= 65 && c < 91)) return 1;
